Big changes and instantiated code
Adjusted code so it looks neater and gave a more in depth description of certain actions that were occurring within the game.
adjusted player movement and jumping to make it feel more like a game.
added box collider to player character to ensure that it can only jump once and not multiple times.
adjusted gravity scale, angular drag, interpolation and collision detection to continuous to make the character feel better to control.