Added Main Menu
Added a main menu to the game, this is to set myself up for tomorrows coding where i will be adding a play button to allow for the player to actually access the game.
Also some files that weren't saved on unity before applying last commit, so they have been included in here too.